qml: Fixes to handling FPS changes, adjust time on FPS change
Clear the old subtitle when loading new one that requires manually set
FPS. Set the initial FPS to be empty string.

Adjust the playback time when FPS is changed to be in the correct frame.
Works only when the new time is not exceeding total time, in this case
it is reset to 0. Removed the notify on that one.