engine: simplify time manipulation, fix offset use and logic flaws
Use simpler approach for manipulating time. This might make it easier
to change later if needed.

Fix the use of offset, use it always in the calculation by introducing
an initial delay state.

Fix logic flaws in the engine code. For example, always set new
iPrevEndTime when new time is set.